refactor(foundation): reduce module host complexity
